Optimization of hot paths (#3970)
* Keep foldedComponentIds as a string

Seems the only use just joins the contents on every read, so we can do it on write instead.

* Build className string manually

Items are added to the array only to be filtered out again and then the remaining ones get joined. This performs better.

* Use Set for domElements

More accurately represents the collection and gives O(1) lookup performance.

* Read value once

* Remove unnecessary casting & array check

`flatten` always returns an array, and `join` doesn't seem to care what type it is.

* Simplify objToCssArray

* Make names a local variable

It's not accessed outside of this function and is always reset at the start.

* Build string directly instead of joining array elements

* Add util function for joining string arrays

`[...].join()` for an array of strings only becomes more performant with many thousands of items.
With 10 strings, this function is ~1.5x faster in Chrome and ~3x faster in Firefox.

* Optimize hyphenateStyleName

Just replacing regex with manual iteration.
This could be several times faster if we checked for the "ms-" prefix within the loop, but it makes the function pretty unreadable.